tag: updating var type to get products by category

+++ src/main/java/com/shopspace/shopspacewebservice/client/ProductClient.java
@@ -13,5 +13,5 @@ public interface ProductClient {
     @GetMapping("/limit-data")
     List<ProductDTO> getLastProductsWithLimit(@RequestParam Integer page, @RequestParam Integer size);
     @GetMapping("/by-category")
-    List<ProductDTO> getProductsByCategoryWithLimit(@RequestParam Integer idCategory, @RequestParam Integer page, @RequestParam Integer size);
+    List<ProductDTO> getProductsByCategoryWithLimit(@RequestParam Long idCategory, @RequestParam Integer page, @RequestParam Integer size);
 }
